IVHD/requirements.txt
matplotlib==3.3.2
networkx==2.8
numba==0.55.1
numpy==1.21.2
pandas==1.3.3
scikit_learn==1.0.2
scipy==1.7.2
seaborn==0.11.2
faiss-cpu
